+++ android/src/main/kotlin/com/liasica/a11y_service/extensions.kt
@@ -6,9 +6,7 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import android.provider.Settings
-import android.view.accessibility.AccessibilityEvent
 import android.view.accessibility.AccessibilityNodeInfo
-import androidx.annotation.RequiresApi
 
 fun Context.performAction(action: Int) = require().performGlobalAction(action)